core: replace old project files
