--- RandomPrayerUse.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32407.337
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RandomPrayerUse", "RandomPrayerUse\RandomPrayerUse.csproj", "{CD481ED9-7633-4AC7-92F8-CE5C019076CF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CD481ED9-7633-4AC7-92F8-CE5C019076CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CD481ED9-7633-4AC7-92F8-CE5C019076CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CD481ED9-7633-4AC7-92F8-CE5C019076CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CD481ED9-7633-4AC7-92F8-CE5C019076CF}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {125F6C45-A7CA-4C72-8F88-876FD80808F8}
-	EndGlobalSection
-EndGlobal
