--- RandomPrayerUse/PrayerConfig.cs
@@ -1,16 +1,0 @@
-﻿
-namespace RandomPrayerUse
-{
-    [System.Serializable]
-    public class PrayerConfig
-    {
-        public bool OnlyShuffleOwnedPrayers { get; set; }
-        public bool RemoveMirabras { get; set; }
-
-        public PrayerConfig()
-        {
-            OnlyShuffleOwnedPrayers = false;
-            RemoveMirabras = true;
-        }
-    }
-}
