--- resources/config/Random Prayer.cfg
@@ -1,4 +1,0 @@
-{
-  "OnlyShuffleOwnedPrayers": false,
-  "RemoveMirabras": true
-}
